Split out bell component
