+++ src/prompts/ui.tsx
@@ -1,6 +1,3 @@
-import './bell.scss';
-import './overlay.scss';
-
 import { Component, Fragment, h } from 'preact';
 
 import { PromptConfig } from '../core';
@@ -8,18 +5,20 @@ import { PromptManagerState } from '.';
 import { PushSubscriptionState } from '../core/push';
 import { createPortal } from 'preact/compat';
 import { getBrowserName } from '../core/utils';
+import { Bell } from './bell';
+import { Alert } from './alert';
 
-const DEFAULT_SUBSCRIBE_LABEL = 'Subscribe for notifications';
+export const DEFAULT_SUBSCRIBE_LABEL = 'Subscribe for notifications';
 
-function inversePosition(pos: string): 'left' | 'right' {
+export function inversePosition(pos: string): 'left' | 'right' {
     return pos.indexOf('left') > -1 ? 'right' : 'left';
 }
 
-function isMobile(): boolean {
+export function isMobile(): boolean {
     return /android|iPhone|iPad|iPod|mobile/i.test(navigator.userAgent);
 }
 
-interface PromptUiProps {
+export interface PromptUiProps {
     config: PromptConfig;
     subscriptionState: PushSubscriptionState;
     promptManagerState: PromptManagerState;
@@ -31,57 +30,13 @@ interface TooltipProps {
     position: 'left' | 'right';
 }
 
-class Tooltip extends Component<TooltipProps, never> {
+export class Tooltip extends Component<TooltipProps, never> {
     render() {
         return (
             <div
                 class={`kumulos-tooltip kumulos-tooltip-${this.props.position}`}
             >
                 {this.props.children}
-            </div>
-        );
-    }
-}
-
-class Bell extends Component<PromptUiProps, never> {
-    onRequestNativePrompt = () => {
-        this.props.onPromptAccepted(this.props.config);
-    };
-
-    render() {
-        const classes = `kumulos-prompt kumulos-prompt-${this.props.promptManagerState} kumulos-bell-container kumulos-bell-container-${this.props.config.position}`;
-        const tooltipPos = inversePosition(this.props.config.position);
-        const bgColor = this.props.config.colors?.bell?.bg;
-        const fgColor = this.props.config.colors?.bell?.fg;
-
-        const bellStyle = {
-            borderColor: fgColor,
-            backgroundColor: bgColor
-        };
-
-        return (
-            <div class={classes}>
-                <div class="kumulos-bell-inner kumulos-tooltip-parent">
-                    <div
-                        class="kumulos-bell"
-                        onClick={this.onRequestNativePrompt}
-                        style={bellStyle as any}
-                    >
-                        <svg
-                            viewBox="0 0 24 24"
-                            xmlns="http://www.w3.org/2000/svg"
-                        >
-                            <path
-                                d="M12 22c1.1 0 2-.9 2-2h-4c0 1.1.89 2 2 2zm6-6v-5c0-3.07-1.64-5.64-4.5-6.32V4c0-.83-.67-1.5-1.5-1.5s-1.5.67-1.5 1.5v.68C7.63 5.36 6 7.92 6 11v5l-2 2v1h16v-1l-2-2z"
-                                fill={fgColor}
-                            />
-                        </svg>
-                    </div>
-                    <Tooltip position={tooltipPos}>
-                        {this.props.config.labels?.tooltip?.subscribe ??
-                            DEFAULT_SUBSCRIBE_LABEL}
-                    </Tooltip>
-                </div>
             </div>
         );
     }
@@ -265,6 +220,16 @@ export default class Ui extends Component<UiProps, UiState> {
                         onPromptDeclined={this.props.onPromptDeclined}
                     />
                 );
+            case 'alert':
+                return (
+                    <Alert
+                        config={prompt}
+                        subscriptionState={this.props.subscriptionState}
+                        promptManagerState={this.props.promptManagerState}
+                        onPromptAccepted={this.props.onPromptAccepted}
+                        onPromptDeclined={this.props.onPromptDeclined}
+                    />
+                )
             default:
                 return null;
         }
